example_app/requirements.txt
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
faiss-cpu==1.7.4
python-dotenv==1.0.1
PyYAML==6.0.1
